+++ di/src/main/kotlin/team/applemango/runnerbe/di/RunningItemUseCaseModule.kt
@@ -2,12 +2,12 @@
  * RunnerBe © 2022 Team AppleMango. all rights reserved.
  * RunnerBe license is under the MIT.
  *
- * [UseCaseModule.kt] created by Ji Sungbin on 22. 3. 2. 오후 9:30
+ * [RunningItemUseCaseModule.kt] created by Ji Sungbin on 22. 3. 20. 오전 12:39
  *
  * Please see: https://github.com/applemango-runnerbe/RunnerBe-Android/blob/main/LICENSE.
  */
 
-package team.applemango.runnerbe.di.module
+package team.applemango.runnerbe.di
 
 import dagger.Module
 import dagger.Provides
@@ -16,19 +16,18 @@ import dagger.hilt.android.components.ViewModelComponent
 import dagger.hilt.android.scopes.ViewModelScoped
 import team.applemango.runnerbe.domain.runningitem.repository.RunningItemRepository
 import team.applemango.runnerbe.domain.runningitem.usecase.LoadRunningItemsUseCase
-import team.applemango.runnerbe.domain.user.repository.UserRepository
-import team.applemango.runnerbe.domain.user.usecase.UpdateBookmarkItemUseCase
+import team.applemango.runnerbe.domain.runningitem.usecase.WriteRunningItemUseCase
 
 @Module
 @InstallIn(ViewModelComponent::class)
-internal object UseCaseModule {
-    @Provides
-    @ViewModelScoped
-    fun provideUpdateBookmarkItemUseCase(repo: UserRepository): UpdateBookmarkItemUseCase =
-        UpdateBookmarkItemUseCase(repo)
-
+object RunningItemUseCaseModule {
     @Provides
     @ViewModelScoped
     fun provideLoadRunningItemsUseCase(repo: RunningItemRepository): LoadRunningItemsUseCase =
         LoadRunningItemsUseCase(repo)
+
+    @Provides
+    @ViewModelScoped
+    fun provideWriteRunningItemUseCase(repo: RunningItemRepository): WriteRunningItemUseCase =
+        WriteRunningItemUseCase(repo)
 }
